Ease MPI involution
The idea behind this change is to send directly contigous rows, that
are easier to access, so the implementation supports it by creating
an initial greater abstraction to split the available rows for each
column and directly store work size and save addresses.
This allows to create smaller wrapped messages of "contiguous" rows
(they are trully contiguous only at first step) and therefore it
makes the program faster.